Support option inheritance for subcommand classes
This implements support for option inheritance in the case of passing
a Clamp::Command subclass as a parameter to #subcommand, vs. as a block.

The key change is in subcommand/declaration.rb, with the rest largely
being fallout to make that work as expected.  Of note, Attributes now
persist any block they're given for later use when defining an attribute
writer method.